clean up the contents of the generate folder
