+++ generate/combyne/filters/has_return_type.js
