+++ generate/combyne/filters/has_returns.js
