+++ generate/combyne/filters/is_double_pointer.js
