+++ generate/combyne/filters/is_pointer.js
