+++ generate/combyne/filters/returns_info.js
