+++ generate/combyne/filters/upper.js
