+++ generate/combyne/manual/include/functions/copy.h
