+++ generate/combyne/manual/include/macros.h
