+++ generate/combyne/manual/src/functions/copy.cc
