+++ generate/combyne/manual/src/wrapper.cc
