+++ generate/combyne/partials/async_function.cc
