+++ generate/combyne/partials/doc.cc
