+++ generate/combyne/partials/guard_arguments.cc
