+++ generate/combyne/partials/sync_function.cc
