+++ generate/combyne/templates/binding.gyp
