+++ generate/combyne/templates/class_content.cc
