+++ generate/combyne/templates/class_header.h
