+++ generate/combyne/templates/enums.js
