+++ generate/combyne/templates/nodegit.cc
