+++ generate/combyne/templates/struct_content.cc
