+++ generate/combyne/templates/struct_header.h
